pyproject.toml
[project]
name = "nut65-pipboy"
version = "0.1.0"
description = "Fallout Pip-Boy themed GUI for Weikav NUT65 keyboard LED control"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "hidapi",
    "pywebview",
    "psutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
